pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.force-include]
"assets" = "tygenie/assets"

[project]
name = "tygenie"
license = { file = "LICENSE" }
version = "1.6.2"
authors = [
  { name = "Pierre-Samuel Le Stang", email = "pierre-samuel.le-stang@corp.ovh.com" },
]
description = "Tygenie is an Opsgenie TUI application"
readme = "README.md"
dependencies = [
  "urllib3==2.2.2",
  "attrs==24.2.0",
  "desktop_notifier==5.0.1",
  "httpx==0.27.2",
  "markdownify==0.13.1",
  "pendulum==3.0.0",
  "python_dateutil==2.9.0.post0",
  "rich==13.9.4",
  "textual[syntax]==0.87.1",
  "udatetime==0.0.17",
]
requires-python = ">=3.11"

[project.scripts]
tygenie = "tygenie.app:main"

[project.urls]
"Homepage" = "https://github.com/ovh/tygenie"
